+++ lib/model/asset.dart
@@ -1,4 +1,7 @@
+// ignore_for_file: depend_on_referenced_packages
+
 import 'package:json_annotation/json_annotation.dart';
+import 'package:monitor_app/db/models/assets_db.dart';
 
 part 'asset.g.dart';
 
@@ -8,7 +11,7 @@ class Asset {
   final String section;
   final String category;
   final String description;
-  final String url;
+  late String url;
   final String createdDate;
   late String? lastModified;
   late bool isPassed;
@@ -30,4 +33,13 @@ class Asset {
 
   Map<String, dynamic> toJson() => _$AssetToJson(this);
   factory Asset.fromJson(Map<String, dynamic> json) => _$AssetFromJson(json);
+
+  factory Asset.fromAssetDB(AssetsDB assetDB) => Asset(
+      section: assetDB.section,
+      id: assetDB.id,
+      category: assetDB.category,
+      description: assetDB.description,
+      url: assetDB.url,
+      createdDate: assetDB.createdDate,
+      orderIndex: assetDB.orderIndex);
 }