+++ lib/screen/album_screen.dart
@@ -102,9 +102,7 @@ class _AlbumScreenState extends State<AlbumScreen> {
                         )
                             .then((value) {
                           if (value != null) {
-                            if (value) {
-                              Navigator.of(context).pop();
-                            }
+                            Navigator.pop(context, value);
                           }
                         });
                         refreshAlreadyCapturedImages();