+++ lib/screen/detail_task_screen.dart
@@ -1,8 +1,11 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:flutter/material.dart';
-import 'package:monitor_app/components/task_item_card.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:monitor_app/screen/components/task_item_card.dart';
+import 'package:monitor_app/controller/app_provider.dart';
 import 'package:monitor_app/model/asset.dart';
-class DetailTaskScreen extends StatefulWidget {
+
+class DetailTaskScreen extends ConsumerStatefulWidget {
   final String title;
   // final List<MasterAsset> masterAsset;
   final List<Asset> assets;
@@ -14,10 +17,10 @@ class DetailTaskScreen extends StatefulWidget {
   }) : super(key: key);
 
   @override
-  State<DetailTaskScreen> createState() => _DetailTaskScreenState();
+  ConsumerState<DetailTaskScreen> createState() => _DetailTaskScreenState();
 }
 
-class _DetailTaskScreenState extends State<DetailTaskScreen> {
+class _DetailTaskScreenState extends ConsumerState<DetailTaskScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -29,22 +32,33 @@ class _DetailTaskScreenState extends State<DetailTaskScreen> {
   }
 
   Widget _buildBody() {
+    final state = ref.watch(taskProvider);
     return Column(
       children: [
         Expanded(
           child: Container(
             color: const Color(0xFFEAEEF2),
             padding: const EdgeInsets.all(10),
-            child: GridView.builder(
-              itemCount: widget.assets.length,
-              gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-                crossAxisCount: 2,
-                crossAxisSpacing: 8,
-                mainAxisSpacing: 8,
-                childAspectRatio: 0.5,
-              ),
-              itemBuilder: (context, index) {
-                return TaskItemCard(asset: widget.assets[index]);
+            child: Consumer(
+              builder: (context, ref, child) {
+                if (state != null) {
+                  return GridView.builder(
+                    itemCount: widget.assets.length,
+                    gridDelegate:
+                        const SliverGridDelegateWithFixedCrossAxisCount(
+                      crossAxisCount: 2,
+                      crossAxisSpacing: 8,
+                      mainAxisSpacing: 8,
+                      childAspectRatio: 0.46,
+                    ),
+                    itemBuilder: (context, index) {
+                      return TaskItemCard(
+                          asset: widget.assets[index], task: state);
+                    },
+                  );
+                } else {
+                  return Container();
+                }
               },
             ),
           ),